add：PI-32972 上报事件中增加信息
