+++ growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/change/ViewChangeProvider.java
@@ -21,8 +21,8 @@ import android.view.View;
 import android.widget.EditText;
 
 import com.growingio.android.sdk.autotrack.Autotracker;
-import com.growingio.android.sdk.autotrack.events.AutotrackEventType;
-import com.growingio.android.sdk.autotrack.events.ViewElementEvent;
+import com.growingio.android.sdk.track.events.AutotrackEventType;
+import com.growingio.android.sdk.track.events.ViewElementEvent;
 import com.growingio.android.sdk.autotrack.page.Page;
 import com.growingio.android.sdk.autotrack.page.PageProvider;
 import com.growingio.android.sdk.autotrack.view.OnViewStateChangedListener;
