+++ growingio-autotracker-core/src/main/java/com/growingio/android/sdk/autotrack/hybrid/event/HybridPageEvent.java
@@ -16,7 +16,7 @@
 
 package com.growingio.android.sdk.autotrack.hybrid.event;
 
-import com.growingio.android.sdk.autotrack.events.PageEvent;
+import com.growingio.android.sdk.track.events.PageEvent;
 
 import org.json.JSONException;
 import org.json.JSONObject;
