+++ growingio-tracker-core/src/main/java/com/growingio/android/sdk/track/events/PageAttributesEvent.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.growingio.android.sdk.autotrack.events;
+package com.growingio.android.sdk.track.events;
 
 import com.growingio.android.sdk.track.events.base.BaseAttributesEvent;
